requirements.txt
pygame==2.6.1
numpy>=1.21.0
scipy>=1.7.0
pyOpenBCI>=0.12.0
pyserial>=3.0
xmltodict>=0.13.0
